Convert vigilante suicide to a killer tag
This allows transition_day to correctly determine the vigilante's
location and for us to still be able to suppress protections. Formerly,
the add_dying() would circumvent much of transition_day's logic because
they were already treated as dead, which led to the situation where a
wolf attacking a vigilante who guesses wrong will play a "target not
home" message due to said vigilante already being dead (and therefore
not having a location).