Address issue with loggedin event
This might be anope-specific, but I didn't test that; shouldn't impact
most things anyway if it's generic. The loggedin event seems to give us
our uncloaked nick!ident@host, even if we were already given a cloak via
a prior hosthidden event. This causes us to not recognize ourselves when
we join channels, making the bot unresponsive.

This particular path typically only happens if SASL is not in use; as if
SASL is used we'll get loggedin before any hosthidden replies, and
therefore have the correct host associated with the bot user.