Vampire bite no longer silences the target
The intent behind this was so that while the person was out resting
during the day to recover from the blood loss, they couldn't do day
actions (such as shoot, id, ...). However, silence lasts until the end
of night, which is not desirable in this particular instance.

Rather than trying to make individual silence instances expire at
different times based on the role that set it, it's easier to just
remove the silence aspect from vampire bite.