(BREAKING CHANGE) Make likelihoods configurable
Rather than encoding the likelihood of a mode as part of the game_mode
decorator, it is now present in botconfig.yml. The defaults have been
migrated over, however any custom modes will require manual work.

If your custom mode simply re-defined a built-in mode to change the
likelihood, delete the custom mode file and customize the likelihood via
botconfig.yml instead.

If you have actual custom modes, you must remove the likelihood kwarg
from the game_mode decorator, and then add a section to your
botconfig.yml to define the mode's likelihood, like so:

gameplay:
  <mode name>:
    weight: <likelihood>

The default is 0 for custom modes that are missing from botconfig, i.e.
the modes will be majority-vote only. Failure to remove the likelihood
kwarg will result in the bot failing to start. This is a breaking
change.

Additionally, for sleepy and boreal some additional settings became
configurable in botconfig.yml which can be used to tweak the difficulty
of those modes.

Finally, the configuration page autogeneration for the wiki should now
properly handle union types.