Don't try to add users to a channel named "*"
WHO and WHOX replies use * as a placeholder for the channel parameter if
we have no visibility into a channel for the specified user. Internally,
we also have a fake channel called * aliased to channels.Dummy for use
in testing. The combination meant that sometimes users got added to the
fake Dummy channel, which could break user tracking later on if they
quit. Fix WHO and WHOX reply handling to ignore channel parameters
containing * so this doesn't happen.