Unrestrict vigilante kill in pactbreaker
They still die if they target incorrectly, but they no longer require
evidence in order to use the !kill command. This should in turn make it
a bit more likely that wolf dies before vigilante dies/turns, which in
turn makes it a bit easier for vigilante (and vampire) to win.